Refactored common styled.js into smaller files for individual types
