+++ src/features/InfoPage/AuthorPanel/index.js
@@ -1,10 +1,8 @@
 import { SmallHeading } from "./styled";
-import {
-    CommonContainer,
-    GridWrapper,
-    MinorHeading,
-    UserAvatar
-} from "../../../common/styled";
+import { CommonContainer } from "../../../common/Containers/styled";
+import { MinorHeading } from "../../../common/Texts/styled";
+import { UserAvatar } from "../../../common/Elements/styled";
+import { GridWrapper } from "../../../common/Wrappers/styled";
 
 const AuthorPanel = () => (
     <CommonContainer>
