+++ src/features/MainPage/UsersList/Actions/Idle/index.js
@@ -1,8 +1,6 @@
 import { MapIcon } from "../styled";
-import {
-    CommonContainer,
-    CommonParagraph
-} from "../../../../../common/styled";
+import { CommonContainer } from "../../../../../common/Containers/styled";
+import { CommonParagraph } from "../../../../../common/Texts/styled";
 
 const Idle = () => (
     <CommonContainer $action>
